move violinRenderer brush test from violin plot integration spec to a dom unit test
